* get rid of loops on git repositories with merges
in case of the following topology:
   H
B1   B2
   R
`git log B1..H` always has B2 when `git log B2..H` always has B1
we end up in a loop. let's use date of the current revision to
to cut of anything older. In this case some commits in branches
wouldn't be tested